backend/share/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autoflake==1.4
black==20.8b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
djangorestframework==3.12.2
drf-yasg==1.20.0
flake8==3.8.4
idna==2.10
inflection==0.5.1
invoke==1.4.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==20.7
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
